Clean .env file before deploying
When running in a pipeline, the .env file needs to be renegerated or you
risk deploying with settings from the previous stage. Adding `rm_env` as
a prerequisite to `deploy` will regenerate the .env file, but only when
being run in a GoCD environment (by detecting the presence of an
environment variable present in all GoCD pipelines).