Zip node_modules after downloading
npm/yarn creates a lot of files when downloading dependencies, which can
be slow on filesystems with a lot of IO latency like EFS. This change
zips node_modules after downloading them, which allows you to upload a
single file as an artifact. This also has the benefit of greatly
reducing the artifact size.